requirements.txt
marker-pdf
python-dotenv
google-generativeai
google-genai
markdown
python-pptx
requests
pdfminer.six
weasyprint